+++ index.js
@@ -6,7 +6,7 @@ const SepsadSecurityConst = require('./sepsadSecurityConst');
 
 function mySepsadSecurityPlatform(log, config, api) {
   if (!config) {
-    log('No configuration found for homebridge-sepsadSecurity');
+    log('No configuration found for homebridge-sepsadsecurity');
     return;
   }
 
@@ -39,6 +39,10 @@ function mySepsadSecurityPlatform(log, config, api) {
   this.sepsadSecurityAPI = new SepsadSecurityAPI(log, this);
 
   this.loaded = false;
+  this.tempLoaded = false;
+
+  this._confirmedAccessories = [];
+  this._confirmedServices = [];
 
   this.api
     .on(
@@ -55,7 +59,7 @@ function mySepsadSecurityPlatform(log, config, api) {
         if (this.cleanCache) {
           this.log('WARNING - Removing Accessories');
           this.api.unregisterPlatformAccessories(
-            'homebridge-sepsadSecurity',
+            'homebridge-sepsadsecurity',
             'SepsadSecurity',
             this.foundAccessories
           );
@@ -73,7 +77,7 @@ module.exports = function (homebridge) {
   UUIDGen = homebridge.hap.uuid;
   HomebridgeAPI = homebridge;
   homebridge.registerPlatform(
-    'homebridge-sepsadSecurity',
+    'homebridge-sepsadsecurity',
     'SepsadSecurity',
     mySepsadSecurityPlatform,
     true
@@ -96,10 +100,60 @@ mySepsadSecurityPlatform.prototype = {
     // disconnecting ?
   },
 
+  //Cleaning methods
+  cleanPlatform: function () {
+    this.cleanAccessories();
+    this.cleanServices();
+  },
+
+  cleanAccessories: function () {
+    //cleaning accessories
+    let accstoRemove = [];
+    for (let acc of this.foundAccessories) {
+      if (!this._confirmedAccessories.find((x) => x.UUID == acc.UUID)) {
+        accstoRemove.push(acc);
+        this.log('WARNING - Accessory will be Removed ' + acc.UUID + '/' + acc.displayName);
+      }
+    }
+
+    if (accstoRemove.length > 0)
+      this.api.unregisterPlatformAccessories(
+        'homebridge-sepsadsecurity',
+        'SepsadSecurity',
+        accstoRemove
+      );
+  },
+
+  cleanServices: function () {
+    //cleaning services
+    for (let acc of this.foundAccessories) {
+      let servicestoRemove = [];
+      for (let serv of acc.services) {
+        if (
+          serv.subtype !== undefined &&
+          !this._confirmedServices.find((x) => x.UUID == serv.UUID && x.subtype == serv.subtype)
+        ) {
+          servicestoRemove.push(serv);
+        }
+      }
+      for (let servToDel of servicestoRemove) {
+        this.log(
+          'WARNING - Service Removed' +
+            servToDel.UUID +
+            '/' +
+            servToDel.subtype +
+            '/' +
+            servToDel.displayName
+        );
+        acc.removeService(servToDel);
+      }
+    }
+  },
+
   discoverSecuritySystem: function () {
     this.sepsadSecurityAPI.on('securitySystemRefreshError', () => {
       if (this.timerID == undefined) {
-        this.log('ERROR - discoverSecuritySystem - will retry in 1 minute');
+        this.log('ERROR - securitySystemRefreshError - will retry in 1 minute');
         setTimeout(() => {
           this.sepsadSecurityAPI.getSecuritySystem();
         }, 60000);
@@ -108,6 +162,10 @@ mySepsadSecurityPlatform.prototype = {
 
     this.sepsadSecurityAPI.on('securitySystemRefreshed', () => {
       this.log.debug('INFO - securitySystemRefreshed event');
+      this.log.debug(
+        'INFO - SecuritySystem : ' + JSON.stringify(this.sepsadSecurityAPI.securitySystem)
+      );
+
       if (!this.loaded) {
         this.loadSecuritySystem();
       } else {
@@ -115,19 +173,41 @@ mySepsadSecurityPlatform.prototype = {
       }
     });
 
+    this.sepsadSecurityAPI.on('securitySystemTemperatureRefreshError', () => {
+      if (this.timerID == undefined) {
+        this.log('ERROR - securitySystemTemperatureRefreshError - will retry in 1 minute');
+        setTimeout(() => {
+          this.sepsadSecurityAPI.getSecuritySystem();
+        }, 60000);
+      }
+    });
+
+    this.sepsadSecurityAPI.on('securitySystemTemperatureRefreshed', () => {
+      this.log.debug('INFO - securitySystemTemperatureRefreshed event');
+      this.log.debug(
+        'INFO - securitySystemTemperature : ' +
+          JSON.stringify(this.sepsadSecurityAPI.securitySystem.temperatureInfo)
+      );
+
+      if (!this.tempLoaded) {
+        this.createTemperatureSensorsAccessories();
+      } else {
+        this.updateTemperature();
+      }
+    });
+
     this.sepsadSecurityAPI.getSecuritySystem();
   },
 
-  loadSecuritySystem() {
-    if (this.sepsadSecurityAPI.securitySystem) {
-      this.log.debug(
-        'INFO - SecuritySystem : ' + JSON.stringify(this.sepsadSecurityAPI.securitySystem)
-      );
+  createSecuritySystemAccessory() {
+    if (this.sepsadSecurityAPI.securitySystem.security) {
       let securitySystemName = this.sepsadSecurityAPI.securitySystem.name;
       let securitySystemModel = this.sepsadSecurityAPI.securitySystem.model;
       let securitySystemSeriaNumber = this.sepsadSecurityAPI.securitySystem.id;
 
-      let uuid = UUIDGen.generate(securitySystemName);
+      this.log('INFO - Discovered SecuritySystem : ' + securitySystemName);
+
+      let uuid = UUIDGen.generate(securitySystemName + securitySystemSeriaNumber);
       let mySecuritySystemAccessory = this.foundAccessories.find((x) => x.UUID == uuid);
 
       if (!mySecuritySystemAccessory) {
@@ -144,7 +224,7 @@ mySepsadSecurityPlatform.prototype = {
           .setCharacteristic(Characteristic.Model, mySecuritySystemAccessory.model)
           .setCharacteristic(Characteristic.SerialNumber, mySecuritySystemAccessory.serialNumber);
 
-        this.api.registerPlatformAccessories('homebridge-sepsadSecurity', 'SepsadSecurity', [
+        this.api.registerPlatformAccessories('homebridge-sepsadsecurity', 'SepsadSecurity', [
           mySecuritySystemAccessory,
         ]);
 
@@ -172,6 +252,9 @@ mySepsadSecurityPlatform.prototype = {
       this.bindSecuritySystemCurrentStateCharacteristic(HKSecurityService);
       this.bindSecuritySystemTargetStateCharacteristic(HKSecurityService);
 
+      this._confirmedAccessories.push(mySecuritySystemAccessory);
+      this._confirmedServices.push(HKSecurityService);
+
       // Required Characteristics
       //this.addCharacteristic(Characteristic.SecuritySystemCurrentState);
       //this.addCharacteristic(Characteristic.SecuritySystemTargetState);
@@ -180,6 +263,178 @@ mySepsadSecurityPlatform.prototype = {
       //this.addOptionalCharacteristic(Characteristic.StatusTampered);
       //this.addOptionalCharacteristic(Characteristic.SecuritySystemAlarmType);
       //this.addOptionalCharacteristic(Characteristic.Name);
+    }
+  },
+
+  createSmokeSensorsAccessories() {
+    if (this.sepsadSecurityAPI.securitySystem.fire) {
+      let smokeDetectors = this.sepsadSecurityAPI.securitySystem.fire.smokeDetectors;
+
+      for (let a = 0; a < smokeDetectors.length; a++) {
+        let smokeSensorName = smokeDetectors[a].label;
+        let smokeSensorModel = this.sepsadSecurityAPI.securitySystem.model;
+        let smokeSensorSeriaNumber =
+          this.sepsadSecurityAPI.securitySystem.id + '/' + smokeDetectors[a].id;
+
+        this.log('INFO - Discovered SmokeSensor : ' + smokeSensorName);
+
+        let uuid = UUIDGen.generate(smokeSensorName + smokeSensorSeriaNumber);
+
+        let mySmokeSensorAccessory = this.foundAccessories.find((x) => x.UUID == uuid);
+
+        if (!mySmokeSensorAccessory) {
+          mySmokeSensorAccessory = new Accessory(smokeSensorName, uuid);
+          mySmokeSensorAccessory.name = smokeSensorName;
+          mySmokeSensorAccessory.model = smokeSensorModel;
+          mySmokeSensorAccessory.manufacturer = 'Sepsad/EPS';
+          mySmokeSensorAccessory.serialNumber = smokeSensorSeriaNumber;
+          mySmokeSensorAccessory.smokeSensorID = smokeDetectors[a].id;
+
+          mySmokeSensorAccessory
+            .getService(Service.AccessoryInformation)
+            .setCharacteristic(Characteristic.Manufacturer, mySmokeSensorAccessory.manufacturer)
+            .setCharacteristic(Characteristic.Model, mySmokeSensorAccessory.model)
+            .setCharacteristic(Characteristic.SerialNumber, mySmokeSensorAccessory.serialNumber);
+
+          this.api.registerPlatformAccessories('homebridge-sepsadsecurity', 'SepsadSecurity', [
+            mySmokeSensorAccessory,
+          ]);
+
+          this.foundAccessories.push(mySmokeSensorAccessory);
+        }
+
+        mySmokeSensorAccessory.smokeSensorID = smokeDetectors[a].id;
+        mySmokeSensorAccessory.name = smokeSensorName;
+
+        let HKSmokeSensorService = mySmokeSensorAccessory.getServiceByUUIDAndSubType(
+          smokeSensorName,
+          'SmokeSensorService' + smokeSensorName
+        );
+
+        if (!HKSmokeSensorService) {
+          this.log('INFO - Creating SmokeSensor Service ' + smokeSensorName);
+          HKSmokeSensorService = new Service.SmokeSensor(
+            smokeSensorName,
+            'SmokeSensorService' + smokeSensorName
+          );
+          HKSmokeSensorService.subtype = 'SmokeSensorService' + smokeSensorName;
+          mySmokeSensorAccessory.addService(HKSmokeSensorService);
+        }
+
+        this.bindSmokeDetectedCharacteristic(HKSmokeSensorService);
+        this.bindSmokeStatusActiveCharacteristic(HKSmokeSensorService);
+
+        this._confirmedAccessories.push(mySmokeSensorAccessory);
+        this._confirmedServices.push(HKSmokeSensorService);
+        //this.bindSmokeStatusLowBatteryCharacteristic(HKSmokeSensorService);
+        //this.bindSmokeStatusFaultCharacteristic(HKSmokeSensorService);
+
+        // Required Characteristics
+        //this.addCharacteristic(Characteristic.SmokeDetected);
+
+        // Optional Characteristics
+        //this.addOptionalCharacteristic(Characteristic.StatusActive);
+        //this.addOptionalCharacteristic(Characteristic.StatusFault);
+        //this.addOptionalCharacteristic(Characteristic.StatusTampered);
+        //this.addOptionalCharacteristic(Characteristic.StatusLowBattery);
+        //this.addOptionalCharacteristic(Characteristic.Name);
+      }
+    }
+  },
+
+  createTemperatureSensorsAccessories() {
+    if (this.sepsadSecurityAPI.securitySystem.temperatureInfo) {
+      let tempSensors = this.sepsadSecurityAPI.securitySystem.temperatureInfo;
+      for (let a = 0; a < tempSensors.length; a++) {
+        let tempSensorName = tempSensors[a].label;
+        let tempSensorModel = this.sepsadSecurityAPI.securitySystem.model;
+        let tempSensorSeriaNumber =
+          this.sepsadSecurityAPI.securitySystem.id + '/' + tempSensors[a].id;
+
+        this.log('INFO - Discovered TemperatureSensor : ' + tempSensorName);
+
+        let uuid = UUIDGen.generate(tempSensorName + tempSensorSeriaNumber);
+
+        let myTempSensorAccessory = this.foundAccessories.find((x) => x.UUID == uuid);
+
+        if (!myTempSensorAccessory) {
+          myTempSensorAccessory = new Accessory(tempSensorName, uuid);
+          myTempSensorAccessory.name = tempSensorName;
+          myTempSensorAccessory.model = tempSensorModel;
+          myTempSensorAccessory.manufacturer = 'Sepsad/EPS';
+          myTempSensorAccessory.serialNumber = tempSensorSeriaNumber;
+          myTempSensorAccessory.tempSensorID = tempSensors[a].id;
+
+          myTempSensorAccessory
+            .getService(Service.AccessoryInformation)
+            .setCharacteristic(Characteristic.Manufacturer, myTempSensorAccessory.manufacturer)
+            .setCharacteristic(Characteristic.Model, myTempSensorAccessory.model)
+            .setCharacteristic(Characteristic.SerialNumber, myTempSensorAccessory.serialNumber);
+
+          this.api.registerPlatformAccessories('homebridge-sepsadsecurity', 'SepsadSecurity', [
+            myTempSensorAccessory,
+          ]);
+
+          this.foundAccessories.push(myTempSensorAccessory);
+        }
+
+        myTempSensorAccessory.tempSensorID = tempSensors[a].id;
+        myTempSensorAccessory.name = tempSensorName;
+
+        let HKTempSensorService = myTempSensorAccessory.getServiceByUUIDAndSubType(
+          tempSensorName,
+          'TempSensorService' + tempSensorName
+        );
+
+        if (!HKTempSensorService) {
+          this.log('INFO - Creating TempSensor Service ' + tempSensorName);
+          HKTempSensorService = new Service.TemperatureSensor(
+            tempSensorName,
+            'TempSensorService' + tempSensorName
+          );
+          HKTempSensorService.subtype = 'TempSensorService' + tempSensorName;
+          myTempSensorAccessory.addService(HKTempSensorService);
+        }
+
+        this.bindCurrentTemperatureCharacteristic(HKTempSensorService);
+
+        this._confirmedAccessories.push(myTempSensorAccessory);
+        this._confirmedServices.push(HKTempSensorService);
+
+        //this.bindTempStatusActiveCharacteristic(HKTempSensorService);
+        //this.bindTempStatusLowBatteryCharacteristic(HKTempSensorService);
+        //this.bindTempStatusFaultCharacteristic(HKTempSensorService);
+
+        // Required Characteristics
+        //this.addCharacteristic(Characteristic.CurrentTemperature);
+
+        // Optional Characteristics
+        //this.addOptionalCharacteristic(Characteristic.StatusActive);
+        //this.addOptionalCharacteristic(Characteristic.StatusFault);
+        //this.addOptionalCharacteristic(Characteristic.StatusLowBattery);
+        //this.addOptionalCharacteristic(Characteristic.StatusTampered);
+        //this.addOptionalCharacteristic(Characteristic.Name);
+      }
+
+      this.tempLoaded = true;
+
+      this.cleanPlatform();
+      this.updateTemperature();
+    }
+  },
+
+  loadSecuritySystem() {
+    if (this.sepsadSecurityAPI.securitySystem.systemLastState) {
+      this.createSecuritySystemAccessory();
+      this.createSmokeSensorsAccessories();
+      if (
+        this.sepsadSecurityAPI.securitySystem.temperature &&
+        this.sepsadSecurityAPI.securitySystem.temperature.nbActiveDevices > 0
+      ) {
+        this.sepsadSecurityAPI.getTemperature();
+      } else {
+        this.cleanPlatform();
+      }
 
       this.updateSecuritySystem();
       this.loaded = true;
@@ -200,51 +455,97 @@ mySepsadSecurityPlatform.prototype = {
 
   updateSecuritySystem() {
     for (let a = 0; a < this.foundAccessories.length; a++) {
-      this.log.debug('INFO - refreshing - ' + this.foundAccessories[a].name);
-
-      let securitySystemResult = undefined;
       if (
-        this.sepsadSecurityAPI.securitySystem &&
-        this.sepsadSecurityAPI.securitySystem.id == this.foundAccessories[a].securitySystemID
+        this.foundAccessories[a].securitySystemID &&
+        this.sepsadSecurityAPI.securitySystem.systemLastState
       ) {
-        securitySystemResult = this.sepsadSecurityAPI.securitySystem;
-      }
+        this.log.debug('INFO - refreshing securit System- ' + this.foundAccessories[a].name);
 
-      if (securitySystemResult !== undefined) {
-        this.refreshSecuritySystem(this.foundAccessories[a], securitySystemResult);
-      } else {
-        this.log(
-          'ERROR - updateSecuritySystem - no result for securitySystem - ' +
-            this.foundAccessories[a].name
-        );
+        let securitySystemResult = undefined;
+        if (
+          this.sepsadSecurityAPI.securitySystem &&
+          this.sepsadSecurityAPI.securitySystem.id == this.foundAccessories[a].securitySystemID
+        ) {
+          securitySystemResult = this.sepsadSecurityAPI.securitySystem.systemLastState;
+        }
+
+        if (securitySystemResult !== undefined) {
+          this.refreshSecuritySystem(this.foundAccessories[a], securitySystemResult);
+        } else {
+          this.log(
+            'ERROR - updateSecuritySystem - no result for securitySystem - ' +
+              this.foundAccessories[a].name
+          );
+        }
+      } else if (this.foundAccessories[a].smokeSensorID) {
+        this.log.debug('INFO - refreshing smokeSensor - ' + this.foundAccessories[a].name);
+
+        let smokeResults = this.sepsadSecurityAPI.securitySystem.fire;
+        if (smokeResults !== undefined) {
+          this.refreshSmokeSensor(this.foundAccessories[a], smokeResults);
+        } else {
+          this.log(
+            'ERROR - updateSecuritySystem - no result for smokeSensor - ' +
+              this.foundAccessories[a].name
+          );
+        }
+      }
+    }
+  },
+
+  updateTemperature() {
+    for (let a = 0; a < this.foundAccessories.length; a++) {
+      if (
+        this.foundAccessories[a].tempSensorID &&
+        this.sepsadSecurityAPI.securitySystem.temperatureInfo
+      ) {
+        this.log.debug('INFO - refreshing temp Sensor - ' + this.foundAccessories[a].name);
+
+        let tempResults = this.sepsadSecurityAPI.securitySystem.temperatureInfo;
+        var tempSensorResult = undefined;
+
+        for (let b = 0; b < tempResults.length; b++) {
+          if (tempResults[b].id == this.foundAccessories[a].tempSensorID) {
+            tempSensorResult = tempResults[b];
+            break;
+          }
+        }
+
+        if (tempSensorResult !== undefined) {
+          this.refreshTempSensor(this.foundAccessories[a], tempSensorResult);
+        } else {
+          this.log(
+            'ERROR - updateSecuritySystem - no result for tempSensor - ' +
+              this.foundAccessories[a].name
+          );
+        }
       }
     }
   },
 
   getCurrentSecuritySystemStateCharacteristic: function (service, callback) {
     callback(undefined, service.getCharacteristic(Characteristic.SecuritySystemCurrentState).value);
-
     //no operationInProgress, refresh current state
     if (service.TargetSecuritySystemStateOperationStart == undefined) {
       this.sepsadSecurityAPI.getSecuritySystem();
     }
+
+    // Characteristic.SecuritySystemCurrentState.STAY_ARM = 0;
+    // Characteristic.SecuritySystemCurrentState.AWAY_ARM = 1;
+    // Characteristic.SecuritySystemCurrentState.NIGHT_ARM = 2;
+    // Characteristic.SecuritySystemCurrentState.DISARMED = 3;
+    // Characteristic.SecuritySystemCurrentState.ALARM_TRIGGERED = 4;
   },
 
   getTargetSecuritySystemStateCharacteristic: function (service, callback) {
     callback(undefined, service.getCharacteristic(Characteristic.SecuritySystemTargetState).value);
     //handled through currentState
-  },
 
-  // Characteristic.SecuritySystemCurrentState.STAY_ARM = 0;
-  // Characteristic.SecuritySystemCurrentState.AWAY_ARM = 1;
-  // Characteristic.SecuritySystemCurrentState.NIGHT_ARM = 2;
-  // Characteristic.SecuritySystemCurrentState.DISARMED = 3;
-  // Characteristic.SecuritySystemCurrentState.ALARM_TRIGGERED = 4;
-  // */
-  // Characteristic.SecuritySystemTargetState.STAY_ARM = 0;
-  // Characteristic.SecuritySystemTargetState.AWAY_ARM = 1;
-  // Characteristic.SecuritySystemTargetState.NIGHT_ARM = 2;
-  // Characteristic.SecuritySystemTargetState.DISARM = 3;
+    // Characteristic.SecuritySystemTargetState.STAY_ARM = 0;
+    // Characteristic.SecuritySystemTargetState.AWAY_ARM = 1;
+    // Characteristic.SecuritySystemTargetState.NIGHT_ARM = 2;
+    // Characteristic.SecuritySystemTargetState.DISARM = 3;
+  },
   setTargetSecuritySystemtateCharacteristic: function (service, value, callback) {
     var currentValue = service.getCharacteristic(Characteristic.SecuritySystemTargetState).value;
     var currentState = service.getCharacteristic(Characteristic.SecuritySystemCurrentState).value;
@@ -260,7 +561,17 @@ mySepsadSecurityPlatform.prototype = {
             ' - SecuritySystemCurrentState is ' +
             this.sepsadSecurityAPI.getStateString(currentState)
         );
-        this.sepsadSecurityAPI.activateSecuritySystem(service, function (error) {
+
+        let mode = SepsadSecurityConst.DISABLED;
+        if (value == Characteristic.SecuritySystemTargetState.AWAY_ARM)
+          mode = SepsadSecurityConst.ACTIVATED;
+        else if (
+          value == Characteristic.SecuritySystemTargetState.STAY_ARM ||
+          value == Characteristic.SecuritySystemTargetState.NIGHT_ARM
+        )
+          mode = SepsadSecurityConst.PARTIAL;
+
+        this.sepsadSecurityAPI.activateSecuritySystem(mode, function (error) {
           if (error) {
             that.endSecuritySystemOperation(service);
             that.log.debug(
@@ -300,6 +611,51 @@ mySepsadSecurityPlatform.prototype = {
     }
   },
 
+  getCurrentTemperatureCharacteristic: function (service, callback) {
+    callback(undefined, service.getCharacteristic(Characteristic.CurrentTemperature).value);
+    this.sepsadSecurityAPI.getTemperature();
+  },
+
+  getSmokeDetectedCharacteristic: function (service, callback) {
+    callback(undefined, service.getCharacteristic(Characteristic.SmokeDetected).value);
+
+    //no operationInProgress, refresh current state
+    if (service.TargetSecuritySystemStateOperationStart == undefined) {
+      this.sepsadSecurityAPI.getSecuritySystem();
+    }
+  },
+
+  getSmokeStatusActiveCharacteristic: function (service, callback) {
+    callback(undefined, service.getCharacteristic(Characteristic.StatusActive).value);
+  },
+
+  bindCurrentTemperatureCharacteristic: function (service) {
+    service.getCharacteristic(Characteristic.CurrentTemperature).on(
+      'get',
+      function (callback) {
+        this.getCurrentTemperatureCharacteristic(service, callback);
+      }.bind(this)
+    );
+  },
+
+  bindSmokeDetectedCharacteristic(service) {
+    service.getCharacteristic(Characteristic.SmokeDetected).on(
+      'get',
+      function (callback) {
+        this.getSmokeDetectedCharacteristic(service, callback);
+      }.bind(this)
+    );
+  },
+
+  bindSmokeStatusActiveCharacteristic(service) {
+    service.getCharacteristic(Characteristic.StatusActive).on(
+      'get',
+      function (callback) {
+        this.getSmokeStatusActiveCharacteristic(service, callback);
+      }.bind(this)
+    );
+  },
+
   bindSecuritySystemCurrentStateCharacteristic: function (service) {
     service.getCharacteristic(Characteristic.SecuritySystemCurrentState).on(
       'get',
@@ -310,6 +666,11 @@ mySepsadSecurityPlatform.prototype = {
   },
 
   bindSecuritySystemTargetStateCharacteristic: function (service) {
+    //at startup, set to currentstate to reset
+    service
+      .getCharacteristic(Characteristic.SecuritySystemTargetState)
+      .updateValue(service.getCharacteristic(Characteristic.SecuritySystemCurrentState).value);
+
     service
       .getCharacteristic(Characteristic.SecuritySystemTargetState)
       .on(
@@ -386,11 +747,11 @@ mySepsadSecurityPlatform.prototype = {
   },
 
   operationMode(result) {
-    if (result.state == SepsadSecurityConst.DISABLED) {
+    if (result.securityMode == SepsadSecurityConst.DISABLED) {
       return Characteristic.SecuritySystemCurrentState.DISARMED;
-    } else if (result.state == SepsadSecurityConst.ACTIVATED) {
+    } else if (result.securityMode == SepsadSecurityConst.ACTIVATED) {
       return Characteristic.SecuritySystemCurrentState.AWAY_ARM;
-    } else if (result.state == SepsadSecurityConst.PARTIAL) {
+    } else if (result.securityMode == SepsadSecurityConst.PARTIAL) {
       return Characteristic.SecuritySystemCurrentState.NIGHT_ARM;
     } else {
       return -1;
@@ -458,7 +819,6 @@ mySepsadSecurityPlatform.prototype = {
     var newTargetState = oldTargetState;
 
     //operation has finished or timed out
-
     if (operationInProgressIsFinished || this.endOperation(HKSecurityService, result)) {
       this.endSecuritySystemOperation(HKSecurityService);
       if (!operationInProgressIsFinished) {
@@ -485,6 +845,7 @@ mySepsadSecurityPlatform.prototype = {
           this.sepsadSecurityAPI.getStateString(newTargetState)
       );
       //TargetState before CurrentState
+
       setImmediate(() => {
         HKSecurityService.getCharacteristic(Characteristic.SecuritySystemTargetState).updateValue(
           newTargetState
@@ -507,9 +868,75 @@ mySepsadSecurityPlatform.prototype = {
     }
   },
 
+  refreshTempSensor: function (myTempAccessory, result) {
+    let tempSensorName = myTempAccessory.name;
+
+    let HKTempSensorService = myTempAccessory.getServiceByUUIDAndSubType(
+      tempSensorName,
+      'TempSensorService' + tempSensorName
+    );
+
+    if (!HKTempSensorService) {
+      this.log('Error - refreshTempSensor - ' + tempSensorName + ' - no service found');
+      return;
+    }
+
+    let currentTemp = HKTempSensorService.getCharacteristic(Characteristic.CurrentTemperature)
+      .value;
+    let newTemp = result.temperature;
+
+    if (newTemp != currentTemp)
+      HKTempSensorService.getCharacteristic(Characteristic.CurrentTemperature).updateValue(newTemp);
+  },
+
+  refreshSmokeSensor: function (mySmokeAccessory, result) {
+    let smokeSensorName = mySmokeAccessory.name;
+    let HKSmokeSensorService = mySmokeAccessory.getServiceByUUIDAndSubType(
+      smokeSensorName,
+      'SmokeSensorService' + smokeSensorName
+    );
+
+    if (!HKSmokeSensorService) {
+      this.log('Error - refreshSmokeSensor - ' + smokeSensorName + ' - no service found');
+      return;
+    }
+
+    let currentStatus = HKSmokeSensorService.getCharacteristic(Characteristic.StatusActive);
+    let newStatus = result.fireStatus == 'ON';
+
+    if (currentStatus != newStatus)
+      HKSmokeSensorService.getCharacteristic(Characteristic.StatusActive).updateValue(newStatus);
+
+    var newSmokeStatus = undefined;
+
+    for (let a = 0; a < result.smokeDetectors.length; a++) {
+      if (result.smokeDetectors[a].id == mySmokeAccessory.smokeSensorID) {
+        newSmokeStatus = result.smokeDetectors[a].status != null;
+        break;
+      }
+    }
+
+    if (newSmokeStatus == undefined) {
+      this.log(
+        'Error - refreshSmokeSensor - ' +
+          smokeSensorName +
+          '/' +
+          mySmokeAccessory.smokeSensorID +
+          ' - no data found for smoke detection Status'
+      );
+      return;
+    } else {
+      let currentSmokeStatus = HKSmokeSensorService.getCharacteristic(Characteristic.SmokeDetected);
+      if (currentStatus != newSmokeStatus)
+        HKSmokeSensorService.getCharacteristic(Characteristic.SmokeDetected).updateValue(
+          newSmokeStatus
+        );
+    }
+  },
+
   endOperation(service, result) {
     // TODO CHECK CHECK
-    if (result.state == SepsadSecurityConst.UNKNOWN) return true;
+    if (this.operationMode(result) == SepsadSecurityConst.UNKNOWN) return true;
 
     //timeout
 
@@ -519,7 +946,10 @@ mySepsadSecurityPlatform.prototype = {
     ) {
       let elapsedTime = Date.now() - service.TargetSecuritySystemStateOperationStart;
       this.log.debug(
-        'INFO - CheckTimeout / result : ' + result + ' - elapsedTime : ' + elapsedTime
+        'INFO - CheckTimeout / result : ' +
+          JSON.stringify(result) +
+          ' - elapsedTime : ' +
+          elapsedTime
       );
       if (elapsedTime > this.maxWaitTimeForOperation * 1000) {
         return true;