+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.1.0
+
+- [NEW] smoke and temp sensors
+- [NEW] activation (optionnal)
+- [NEW] cleaning lost devices from cache
+- [FIX] fixes
+
 ## 0.0.3
 
 - [NEW] config schema